requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.31.3
Markdown==3.10
PyJWT==2.10.1
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.2
